+++ resources/views/components/nav.blade.php
@@ -27,6 +27,13 @@
     </x-chief::nav.item>
 @else
     @foreach($items as $navItem)
+
+        @if($loop->first && $title)
+            <div class="text-sm tracking-wider text-grey-500 uppercase" style="padding: 0 0.5rem; margin-bottom: 1rem;">
+                {{ $title }}
+            </div>
+        @endif
+
         <x-chief::nav.item
             label="{{ ucfirst($navItem->label()) }}"
             url="{{ $navItem->url() }}"